add check to exist docker image

+++ automated/convert_rdf.sh
@@ -106,7 +106,9 @@ else
 fi
 
 # docker imageのビルド
-docker rmi rdf-${DATASET} >> ${WORKDIR_LOG}/${YYYYMMDD}_build.log 
+
+image_cnt=`docker images | grep "rdf-${DATASET}" | wc -l`
+test ${image_cnt} -ne 0 && docker rmi rdf-${DATASET} >> ${WORKDIR_LOG}/${YYYYMMDD}_build.log 
 docker build --tag rdf-${DATASET} . >> ${WORKDIR_LOG}/${YYYYMMDD}_build.log
 if [ $? -ne 0 ]; then
   echo "イメージのビルドに失敗しました"
